Create OCP-deploy-Jenkins-build-config
Build configurations define a build process for new container images. There are three types of builds possible - a container image build using a Dockerfile, a Source-to-Image build that uses a specially prepared base image that accepts source code that it can make runnable, and a custom build that can run arbitrary container images as a base and accept the build parameters. Builds run on the cluster and on completion are pushed to the container image registry specified in the "output" section. A build can be triggered via a webhook, when the base image changes, or when a user manually requests a new build be created.

In this section, you create the initial BuildConfig pipeline using one of the two methods listed below.

Read through this sample Build Config Pipeline

Use either of the following methods to create the initial BuildConfig pipeline:

Method 1: Use the command line from your StudentVM host with the OpenShift client:

Create a file with the contents of the BuildConfig pipeline sample above.

Use the oc create -f FILENAME.yaml command to create the BuildConfig pipeline.

Method 2: Use the OpenShift Container Platform web console.

Log in to the OPENTLC shared OpenShift Container Platform web console.

Select the Developer perspective.

Select your pipeline-${GUID}-dev project from the dropdown at the top of the page.

Click the + sign on the top of the page to Import YAML. Paste the Build Config pipeline text that’s above in the text box, and then click Create.

2.6. Test New Pipeline
Using the OpenShift Container Platform web console, make sure that the pipeline was created:

From the left side, Select Builds → Build Configs and then select BC pipeline-demo.

Click Actions → Start Build.